+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/OicrNamingScheme.java
@@ -5,9 +5,11 @@ import org.springframework.web.context.support.SpringBeanAutowiringSupport;
 import uk.ac.bbsrc.tgac.miso.core.data.Library;
 import uk.ac.bbsrc.tgac.miso.core.data.Nameable;
 import uk.ac.bbsrc.tgac.miso.core.data.Sample;
+import uk.ac.bbsrc.tgac.miso.core.data.impl.LibraryAliquot;
 import uk.ac.bbsrc.tgac.miso.core.service.naming.generation.DefaultNameGenerator;
 import uk.ac.bbsrc.tgac.miso.core.service.naming.generation.NameGenerator;
 import uk.ac.bbsrc.tgac.miso.core.service.naming.generation.OicrLibraryAliasGenerator;
+import uk.ac.bbsrc.tgac.miso.core.service.naming.generation.OicrLibraryAliquotAliasGenerator;
 import uk.ac.bbsrc.tgac.miso.core.service.naming.generation.OicrSampleAliasGenerator;
 import uk.ac.bbsrc.tgac.miso.core.service.naming.validation.DefaultNameValidator;
 import uk.ac.bbsrc.tgac.miso.core.service.naming.validation.NameValidator;
@@ -26,6 +28,7 @@ public class OicrNamingScheme extends AbstractNamingScheme {
   private final OicrSampleAliasGenerator sampleAliasGenerator = new OicrSampleAliasGenerator();
   private final OicrLibraryAliasValidator libraryAliasValidator = new OicrLibraryAliasValidator();
   private final OicrLibraryAliasGenerator libraryAliasGenerator = new OicrLibraryAliasGenerator();
+  private final OicrLibraryAliquotAliasGenerator libraryAliquotAliasGenerator = new OicrLibraryAliquotAliasGenerator();
   private final OicrProjectShortNameValidator projectShortNameValidator = new OicrProjectShortNameValidator();
 
   /**
@@ -37,10 +40,11 @@ public class OicrNamingScheme extends AbstractNamingScheme {
     SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(sampleAliasGenerator);
     SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(libraryAliasGenerator);
     SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(libraryAliasValidator);
+    SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(libraryAliquotAliasGenerator);
   }
 
   /**
-   * Sets the SiblingNumberGenerator to use in generating {@link Sample} aliases. Within a Spring context, this will be autowired. This
+   * Sets the SiblingNumberGenerator to use in generating aliases. Within a Spring context, this will be autowired. This
    * method exists for cases where the OicrNamingScheme is not a Spring-managed bean
    * 
    * @param siblingNumberGenerator
@@ -48,6 +52,7 @@ public class OicrNamingScheme extends AbstractNamingScheme {
   public void setSiblingNumberGenerator(SiblingNumberGenerator siblingNumberGenerator) {
     sampleAliasGenerator.setSiblingNumberGenerator(siblingNumberGenerator);
     libraryAliasGenerator.setSiblingNumberGenerator(siblingNumberGenerator);
+    libraryAliquotAliasGenerator.setSiblingNumberGenerator(siblingNumberGenerator);
   }
 
   @Override
@@ -77,6 +82,16 @@ public class OicrNamingScheme extends AbstractNamingScheme {
 
   @Override
   public void setLibraryAliasValidator(NameValidator validator) {
+    throwUnsupported();
+  }
+
+  @Override
+  public void setLibraryAliquotAliasGenerator(NameGenerator<LibraryAliquot> generator) {
+    throwUnsupported();
+  }
+
+  @Override
+  public void setLibraryAliquotAliasValidator(NameValidator validator) {
     throwUnsupported();
   }
 
@@ -117,6 +132,16 @@ public class OicrNamingScheme extends AbstractNamingScheme {
   @Override
   protected NameGenerator<Library> getLibraryAliasGenerator() {
     return libraryAliasGenerator;
+  }
+
+  @Override
+  protected NameValidator getLibraryAliquotAliasValidator() {
+    return libraryAliasValidator;
+  }
+
+  @Override
+  protected NameGenerator<LibraryAliquot> getLibraryAliquotAliasGenerator() {
+    return libraryAliquotAliasGenerator;
   }
 
   @Override